services/feed_service/requirements.txt
anyio==4.12.1
asyncpg==0.31.0
fastapi==0.135.1
httpx==0.28.1
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
redis==5.3.0
SQLAlchemy==2.0.48
uvicorn==0.41.0
